In the previous version we noticed two bottlenecks:
- A lot of String object creation
- A lot of time spent in hashmap.compute

We are creating a string for a location everytime we read a line from
the file. It is unnecessary because we can get away by simply comparing
the bytes. The Java HashMap will still want a copy of the bytes which
is also unnecessary. For a performance critical situation we want more
control over our data structures, so we need to write our own hash
table.

To store data in this hashtable we compute the hashcode of the location
bytes as we read them from the file. There is no point in scanning those
bytes again later to compute the hash code.

If an entry for a location does not exist in the file, then we copy the
bytes and store a new Row object in the table. Otherwise we just need to
update the current Row object in the table.